+++ ReactiveCocoa/ObjC+Messages.swift
@@ -29,7 +29,9 @@ internal let NSMethodSignature: AnyClass = NSClassFromString("NSMethodSignature"
 
 	@objc(getArgument:atIndex:)
 	func copy(to buffer: UnsafeMutableRawPointer?, forArgumentAt index: Int)
-
+	@objc(getReturnValue:)
+	func copyReturnValue(to buffer: UnsafeMutableRawPointer?)
+	
 	func invoke()
 
 	@objc(invocationWithMethodSignature:)
@@ -42,6 +44,8 @@ internal let NSMethodSignature: AnyClass = NSClassFromString("NSMethodSignature"
 
 	@objc(getArgumentTypeAtIndex:)
 	func argumentType(at index: UInt) -> UnsafePointer<CChar>
+	var methodReturnType: UnsafePointer<CChar> { get }
+	var methodReturnLength: UInt { get }
 
 	@objc(signatureWithObjCTypes:)
 	static func signature(withObjCTypes typeEncoding: UnsafePointer<Int8>) -> AnyObject